backend/requirements.txt
flask==2.2.2
python-dotenv==0.19.1
requests==2.26.0
